Revert ":recycle: Refactor: cx 분리"
This reverts commit 5f2ce05d112bf2701208d731e2429f99ed14adf7.
