+++ src/components/TextBox/cx.ts
@@ -1,5 +1,5 @@
 import classNames from 'classnames/bind';
-import style from '../components/Thumbs/index.module.scss';
+import style from './index.module.scss';
 
 const cx = classNames.bind(style);
