+++ src/components/Thumbs/cx.ts
@@ -1,0 +1,6 @@
+import classNames from 'classnames/bind';
+import style from './index.module.scss';
+
+const cx = classNames.bind(style);
+
+export default cx;
